requirements.txt
pandas>=1.5.0
streamlit>=1.28.0
sentence-transformers>=2.2.0
scikit-surprise>=1.1.3
opensearch-py>=2.3.0
boto3>=1.34.0
python-dotenv>=1.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
requests-aws4auth>=1.1.2